// src/interfaces/breakpoints/breakpoints.ts
export interface Devices {
  xs: string;
  sm: string;
  md: string;
  lg: string;
  xl: string;
  xsAndSmaller: string;
  smAndSmaller: string;
  mdAndSmaller: string;
  lgAndSmaller: string;
  xlAndSmaller: string;
  xsAndLarger: string;
  smAndLarger: string;
  mdAndLarger: string;
  lgAndLarger: string;
  xlAndLarger: string;
  mobile: string;
  desktop: string;
}